Add scenarios for household individual telephone capture request Factor reusable event code into helper utility
