+++ acceptance_tests/features/steps/case_events.py
@@ -3,24 +3,15 @@ import functools
 
 from behave import step
 
-from acceptance_tests.utilities.rabbit_helper import start_listening_to_rabbit_queue, store_all_msgs_in_context, \
-    store_first_message_in_context
+from acceptance_tests.utilities.event_helper import get_and_check_case_created_messages
+from acceptance_tests.utilities.rabbit_helper import start_listening_to_rabbit_queue, store_all_msgs_in_context
 from acceptance_tests.utilities.test_case_helper import test_helper
 from config import Config
 
 
 @step("messages are emitted to RH and Action Scheduler with {questionnaire_types} questionnaire types")
 def gather_messages_emitted_with_qids(context, questionnaire_types):
-    context.messages_received = []
-
-    start_listening_to_rabbit_queue(Config.RABBITMQ_RH_OUTBOUND_CASE_QUEUE_TEST,
-                                    functools.partial(store_all_msgs_in_context, context=context,
-                                                      expected_msg_count=len(context.sample_units),
-                                                      type_filter='CASE_CREATED'))
-    test_helper.assertEqual(len(context.messages_received), len(context.sample_units))
-    context.case_created_events = context.messages_received.copy()
-    _test_cases_correct(context)
-    context.messages_received = []
+    get_and_check_case_created_messages(context)
 
     context.expected_uacs_cases = _get_extended_case_created_events_for_uacs(context, questionnaire_types)
     start_listening_to_rabbit_queue(Config.RABBITMQ_RH_OUTBOUND_UAC_QUEUE_TEST,
@@ -77,28 +68,6 @@ def _get_extended_case_created_events_for_uacs(context, questionnaire_types):
     return expected_uacs_cases
 
 
-def _test_cases_correct(context):
-    context.expected_sample_units = context.sample_units.copy()
-
-    for msg in context.case_created_events:
-        _validate_case(msg)
-
-        for index, sample_unit in enumerate(context.expected_sample_units):
-            if _sample_matches_rh_message(sample_unit, msg):
-                del context.expected_sample_units[index]
-                break
-        else:
-            test_helper.fail('Could not find sample unit')
-
-
-def _sample_matches_rh_message(sample_unit, rh_message):
-    return (sample_unit['attributes']['ADDRESS_LINE1'] ==
-            rh_message['payload']['collectionCase']['address']['addressLine1']
-            and sample_unit['attributes']['ADDRESS_LINE2'] ==
-            rh_message['payload']['collectionCase']['address']['addressLine2']
-            and sample_unit['attributes']['REGION'][0] == rh_message['payload']['collectionCase']['address']['region'])
-
-
 def _test_uacs_correct(context):
     test_helper.assertEqual(len(context.messages_received), len(context.expected_uacs_cases))
 
@@ -121,56 +90,3 @@ def _validate_uac_message(parsed_body):
 
 def _uac_message_matches_rh_message(case_created_event, rh_message):
     return case_created_event['payload']['collectionCase']['id'] == rh_message['payload']['uac']['caseId']
-
-
-def _validate_case(parsed_body):
-    test_helper.assertEqual('CENSUS', parsed_body['payload']['collectionCase']['survey'])
-    test_helper.assertEqual('ACTIONABLE', parsed_body['payload']['collectionCase']['state'])
-    test_helper.assertEqual(8, len(parsed_body['payload']['collectionCase']['caseRef']))
-
-
-def get_first_case_created_event(context):
-    start_listening_to_rabbit_queue(Config.RABBITMQ_RH_OUTBOUND_CASE_QUEUE_TEST,
-                                    functools.partial(store_first_message_in_context, context=context,
-                                                      type_filter='CASE_CREATED'))
-    first_case_created_event = context.first_message
-    del context.first_message
-    return first_case_created_event
-
-
-def get_cases_and_uac_event_messages(context):
-    context.messages_received = []
-    start_listening_to_rabbit_queue(Config.RABBITMQ_RH_OUTBOUND_CASE_QUEUE_TEST,
-                                    functools.partial(store_all_msgs_in_context, context=context,
-                                                      expected_msg_count=len(context.sample_units),
-                                                      type_filter='CASE_CREATED'))
-    context.case_created_events = context.messages_received.copy()
-    _test_cases_correct(context)
-    context.messages_received = []
-
-    context.welsh_cases = [case['payload']['collectionCase'] for case in context.case_created_events
-                           if case['payload']['collectionCase']['treatmentCode'].startswith('HH_Q')
-                           and case['payload']['collectionCase']['treatmentCode'].endswith('W')]
-
-    start_listening_to_rabbit_queue(Config.RABBITMQ_RH_OUTBOUND_UAC_QUEUE_TEST,
-                                    functools.partial(store_all_msgs_in_context, context=context,
-                                                      expected_msg_count=get_expected_uac_count(context),
-                                                      type_filter='UAC_UPDATED'))
-    context.uac_created_events = context.messages_received.copy()
-    _test_uacs_updated_correct(context)
-    context.messages_received = []
-
-
-def _test_uacs_updated_correct(context):
-    case_ids = set(case['payload']['collectionCase']['id'] for case in context.case_created_events)
-    test_helper.assertSetEqual(set(uac['payload']['uac']['caseId'] for uac in context.uac_created_events), case_ids)
-    welsh_uac_count = len(tuple(uac_updated_event for uac_updated_event in context.uac_created_events if
-                                uac_updated_event['payload']['uac']['questionnaireId'].startswith('03')))
-    non_welsh_uac_count = len(context.uac_created_events) - welsh_uac_count
-    test_helper.assertEqual(non_welsh_uac_count, len(context.case_created_events))
-
-    test_helper.assertEqual(welsh_uac_count, len(context.welsh_cases))
-
-
-def get_expected_uac_count(context):
-    return len(context.welsh_cases) + len(context.case_created_events)
